Priority-enabled TaskMatcher (#7196)
- Rewrite TaskMatcher to use explicit priority queues, added as
priTaskMatcher.
- Changed taskReader to push all tasks into priTaskMatcher, and loads
more when the number outstanding is too low. Added as priTaskReader.
- Modified Forwarder to work with priTaskMatcher, added as priForwarder.
- Config switch to use old or new matcher, and various small changes to
support that.

Functional tests: This passes all functional tests, but currently only
versioning functional tests flip the switch to use the new matcher (they
exercise matching the most).

Unit tests: Tests for new matcher are not there yet. Other tests will
need to be modified.

- Ability to prioritize uniformly across all pending tasks including
queries+nexus.
- Simpler architecture, no new component between taskReader and
TaskMatcher.
- More separated concerns, e.g. "forwarding" is (mostly) in one place
instead of spread around, "rate limiting" is in one place, etc.
- Easier to understand code, no more nested selects (this is subjective
of course).
- Some behavior improvements, e.g. forwarded backlog tasks don't bounce
back and forth anymore.
- Maybe better performance (after optimizations).

existing functional tests, need to add/update unit tests

lots of new code, may be new bugs

+++ service/matching/forwarder_test.go
@@ -62,7 +62,7 @@ func (t *ForwarderTestSuite) SetupTest() {
 	t.client = matchingservicemock.NewMockMatchingServiceClient(t.controller)
 	t.cfg = &forwarderConfig{
 		ForwarderMaxOutstandingPolls: func() int { return 1 },
-		ForwarderMaxRatePerSecond:    func() int { return 2 },
+		ForwarderMaxRatePerSecond:    func() float64 { return 2 },
 		ForwarderMaxChildrenPerNode:  func() int { return 20 },
 		ForwarderMaxOutstandingTasks: func() int { return 1 },
 	}
